仮完成

+++ HEW_2025/Source/ECS/Systems/Update/Ranking/ResultRankingSystem.h
@@ -22,6 +22,6 @@ class ResultRankingSystem final : public IUpdateSystem
 public:
     void Update(World& world, float dt) override;
 private:
-    int m_posY = 0.7f;
+    float m_posY = 0.7f;
     
 };
